Special Characters
You can’t have special characters in the code. Your original code lead
to the output 'Serve the salade ni\xc3\xa7oise.'
Need to use the unicode representations (e.g. in this case \u00E7).
Rewrote it with that, but commented that one out for now, cause that
causes an error when saving to txt file. For now just using a normal c.
Once we use it for the app, can use the currently out-commented code.